UI_Proj_complex_structure/requirements.txt
streamlit==1.24.0
python-dotenv==1.0.0
google-generativeai==0.7.0
requests==2.32.3
numpy==2.2.3
pandas==2.2.3 